+++ lib/src/string/escape_sequence.dart
@@ -29,7 +29,7 @@ if (state.pos < source.length) {
     state.ok = true;
     {{res}} = v;
   } else if (state.log) {
-    state.error = ErrUnexpected.char(state.pos, Char(source.runeAt(state.pos)));
+    state.error = ErrUnexpected.charAt(state.pos, source);
   }
 } else if (state.log) {
   state.error = ErrUnexpected.eof(state.pos);