+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Fixed minor bug in builder `SkipWhile1`
 - Implemented constructor `ErrUnexpected.charOrEof(this.offset, String source, [int? c])`. Simplified (improved) the generated code of some parsers that parse characters and select the generated error depending on the position of parsing
+- Implemented constructor `ErrUnexpected.charAt(this.offset, String source`
 
 ## 0.14.0